sui_replay_2

Module data_stores

Store implementations for the replay tool.

Modules

data_store

An implementation of the replay interfaces: TransactionStore, EpochStore and ObjectStore backed by the RPC GQL endpoint. Schema in crates/sui-indexer-alt-graphql/schema.graphql. The RPC calls are implemented in gql_queries.rs.

file_system_store

File system implementation of the replay interfaces: TransactionStore, EpochStore, and ObjectStore. Data is persisted on disk under a simple, human-inspectable directory layout.

gql_queries

GQL Queries Interface to the rpc for the gql schema defined in crates\sui-indexer-alt-graphql/schema.graphql. Built in 3 modules: epoch_query, txn_query, object_query. No GQL type escapes this module. From here we return structures defined by the replay tool or bcs encoded data of runtime structures.

in_memory_store

In-memory implementation of the replay interfaces: TransactionStore, EpochStore, and ObjectStore. The InMemoryStore provides fast in-memory lookups.

lru_mem_store

LRU in-memory implementation of the replay interfaces: TransactionStore, EpochStore, and ObjectStore.

read_through_store