sui_replay_2
Module gql_queries

Structs

DynamicFieldArgs

DynamicFieldArgsFields

DynamicFieldAtVersionQuery

Epoch

EpochConnection

EpochData

EpochsArgs

EpochsArgsFields

EpochsQuery

GqlObject

MovePackage

MovePackageConnection

MovePackageVersionFilter

OwnerData

PackageVersionsArgs

PackageVersionsArgsFields

PackageVersionsWithEpochDataQuery

ProtocolConfigs

TransactionBlock

TransactionBlockConnection

TransactionBlockEffects

Functions

dynamic_field_at_version

Return the dynamic field given a parent version. Retrieve the object with at most (and closest to) the version of the parent.

epochs

package_versions_for_replay

Return a (Option<MovePackage>, version, Option<EpochId>) tuple for each version of the package at address. The Option<EpochId> is the epoch in which this package was published.