sui_replay_2::gql_queries
Function package_versions_for_replay

pub async fn package_versions_for_replay(
    client: &Client,
    address: Address,
    pagination_filter: PaginationFilter,
    after_version: Option<u64>,
    before_version: Option<u64>,
) -> Result<Page<(Option<Object>, Version, Option<EpochId>)>, Error>

Return a (Option<MovePackage>, version, Option<EpochId>) tuple for each version of the package at address. The Option<EpochId> is the epoch in which this package was published.