sui_replay_2
Module replay_interface

Logical stores needed by the replay tool. Those stores are loosely modeled after the GQL schema in crates/sui-indexer-alt-graphql/schema.graphql. A TransactionStore is used to retrieve transaction data and effects by digest. An EpochStore is used to retrieve epoch information and protocol configuration. An ObjectStore is used to retrieve objects by their keys, with different query options.

Data is usually retrieved by getting BCS-encoded data rather than navigating the GQL schema. Essentially the code uses the schema to retrieve the data, deserializes it into runtime structures, and then operates on those.

A DataStore with reasonable defaults is provided for convenience (data_store.rs). Other styles of data stores are also provided in data_stores for different use cases.

Structs

EpochData

Epoch data required to reaplay a transaction.

ObjectKey

Query for an object. Specifies an ObjectID and the “rule” to retrieve it.

TransactionInfo

Transaction data with effects and checkpoint required to replay a transaction.

Enums

VersionQuery

Query options for an object. Version request an object at a specific version RootVersion request an object at a given version at most (<=) AtCheckpoint request an object at a given checkpoint. Useful for unknown Version.

Traits

EpochStore

An EpochStore retrieves the epoch data and protocol configuration for a given epoch. Epoch data is collected by an indexer and it is not stored anywhere otherwise. This is a very small amount of information and could conceivably be saved locally and never hit a server.

EpochStoreWriter

Write-back trait for epoch data. Allows storing epoch information.

ObjectStore

The ObjectStore trait is used to retrieve objects by their keys, with different query options.

ObjectStoreWriter

Write-back trait for object data. Allows storing objects and their version mappings based on the query type.

ReadDataStore

Trait combining all read capabilities for a data store

ReadWriteDataStore

Trait combining all read and write capabilities for a data store

SetupStore

A trait to set up the data store. This is used to setup internal state of the data store before starting the replay. At the moment is exclusively used by the FileSystemStore to map network to chain id.

StoreSummary

A reporting trait for data stores to print a usage/summary report.

TransactionStore

A TransactionStore has to be able to retrieve transaction data for a given digest. To replay a transaction the data provided to sui_execution::executor::Executor::execute_transaction_to_effects must be available. Some of that data is not provided by the user. It is naturally available at runtime on a live system and later saved in effects and in the context of a checkpoint.

TransactionStoreWriter

Write-back trait for transaction data. Allows storing transaction data, effects, and checkpoint information.