sui_replay_2::replay_interface
Trait ObjectStore

pub trait ObjectStore {
    // Required method
    fn get_objects(
        &self,
        keys: &[ObjectKey],
    ) -> Result<Vec<Option<(Object, u64)>>, Error>;
}

The ObjectStore trait is used to retrieve objects by their keys, with different query options.

This trait can execute a subset of what is allowed by crates/sui-indexer-alt-graphql/schema.graphql::multiGetObjects. That query likely allows more than what the replay tool needs, which is fairly limited in its usage.

Required Methods

fn get_objects( &self, keys: &[ObjectKey], ) -> Result<Vec<Option<(Object, u64)>>, Error>

Retrieve objects by their keys, with different query options.

If the object is not found, the element in the vector is None. Otherwise each tuple contains:

Object: The object data

u64: The actual version of the object