sui_replay_2::replay_interface
Trait StoreSummary

pub trait StoreSummary {
    // Required method
    fn summary<W: Write>(&self, writer: &mut W) -> Result<()>;
}

A reporting trait for data stores to print a usage/summary report.

Implementors are free to print any relevant statistics or configuration details. The writer allows callers to decide where summaries go (stdout, file, buffers, etc.).

Required Methods

fn summary<W: Write>(&self, writer: &mut W) -> Result<()>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.