sui_replay_2::replay_interface
Trait TransactionStore

pub trait TransactionStore {
    // Required method
    fn transaction_data_and_effects(
        &self,
        tx_digest: &str,
    ) -> Result<(TransactionData, TransactionEffects, u64), Error>;
}

A TransactionStore has to be able to retrieve transaction data for a given digest. To replay a transaction the data in input to sui_execution::executor::Executor::execute_transaction_to_effects has to be provided. Some of that data is not provided by the user. It is naturally available at runtime on a live system and later saved in effects and in the context of a checkpoint.

Required Methods

fn transaction_data_and_effects( &self, tx_digest: &str, ) -> Result<(TransactionData, TransactionEffects, u64), Error>

Given transaction digest, return TransactionData, TransactionEffects and the checkpoint that transaction was executed in.