sui_replay_2
Module tracing

Tracing utilities. Mostly deals with directory/file saving and what gets saved in the trace output.

Functions

get_trace_output_path

Gets the path to store trace output (either the default one ‘./replay’ or user-specified). Upon success, the path will exist in the file system.

save_trace_output

Saves the trace and additional metadata needed to analyze the trace to a subderectory named after the transaction digest.