Crate sui_replay_2

Modules

data_store

An implementation of the replay interfaces: TransactionStore, EpochStore, and ObjectStore. The DataStore is backed by the rpc gql endpoint and the schema defined in crates/sui-indexer-alt-graphql/schema.graphql. The DataStore carries an epoch cache and soon a system package cache that can help when replaying multiple transactions.

execution

Execution module for replay. The call to the executor execute_transaction_to_effects is here and the logic to call it is pretty straightforward. execute_transaction_to_effects requires info from the EpochStore (epoch, protocol config, epoch start timestamp, rgp), from the TransactionStore as in transaction data and effects, and from the ObjectStore for dynamic loads (e.g. synamic fields). This module also contains the traits used by execution to talk to the store (BackingPackageStore, ObjectStore, ChildObjectResolver)

gql_queries

GQL Queries Interface to the rpc for the gql schema defined in crates\sui-indexer-alt-graphql/schema.graphql. Built in 3 modules: epoch_query, txn_query, object_query. No GQL type escapes this module. From here we return structures defined by the replay tool or bcs encoded data of runtime structures.

replay_interface

Logical stores needed by the replay tool. Those stores are loosely modeled after the GQL schema in crates/sui-indexer-alt-graphql/schema.graphql. A TransactionStore is used to retrieve transaction data and effects by digest. An EpochStore is used to retrieve epoch information and protocol configuration. An ObjectStore is used to retrieve objects by their keys, with different query options.

replay_txn

This module contains the logic to use transaction data and effects for loading all the objects and data needed to replay a transaction.

tracing

Tracing utilities. Mostly deals with directory/file saving and what gets saved in the trace output.

Structs

ReplayConfig

Arguments to the replay tool. It allows to replay a single transaction by digest or a file containing multiple digests, one per line. This may evolve to something very different in time and it’s not meant to be stable. The options available are very convenient for the current development cycle.

Enums

Node

Enum around rpc gql endpoints.