sui_replay_2
Module artifacts

Structs

ArtifactManager

Manages artifacts produced by the replay tool. An ArtifactManager is always with respect to a given base path (e.g., the output replay directory for a specific transaction).

ArtifactMember

Enums

Artifact

The types of artifacts that the replay tool knows about and may output.

EncodingType

Encoding types for artifacts that may be output by the replay tool.

Constants

ARTIFACTS

ARTIFACTS_ENCODING_COMPRESSION_EXT

ARTIFACTS_ENCODING_EXT