sui_replay_2

Module execution

Execution module for replay. The call to the executor execute_transaction_to_effects is here and the logic to call it is pretty straightforward. execute_transaction_to_effects requires info from the EpochStore (epoch, protocol config, epoch start timestamp, rgp), from the TransactionStore as in transaction data and effects, and from the ObjectStore for dynamic loads (e.g. dynamic fields). This module also contains the traits used by execution to talk to the store (BackingPackageStore, ObjectStore, ChildObjectResolver)

Structs

ReplayExecutor

TxnContextAndEffects

Functions

execute_transaction_to_effects