sui_replay_2
Module replay_interface

Logical stores needed by the replay tool. Those stores are loosely modeled after the GQL schema in crates/sui-indexer-alt-graphql/schema.graphql. A TransactionStore is used to retrieve transaction data and effects by digest. An EpochStore is used to retrieve epoch information and protocol configuration. An ObjectStore is used to retrieve objects by their keys, with different query options.

Data is usually retrieved by getting the bcs encoded data rather than navigating the GQL schema. Essetially the code uses the schema to retireve the data that is deserilized to runtime structures and work with them.

A DataStore with reasonable defaults is provided for convenience (data_store.rs).

Structs

EpochData

Epoch data required to reaplay a transaction.

ObjectKey

Query for an object. Specifies an ObjectID and the rule to retrieve it.

Enums

VersionQuery

Query options for an object. Version request an object at a specific version RootVersion request an object at a given version at most (<=) AtCheckpoint request an object at a given checkpoint. Useful for unknown Version). ImmutableOrLatest` requests an object assumed to be unversionalbe in the system (e.g. user packages)

Traits

EpochStore

An EpochStore retrieves the epoch data and protocol configuration given an epoch. Epoch data is collected by an indexer and it is not stored anywhere otherwise. This is a very small amount of information and could conceivably be saved locally and never hit a server.

ObjectStore

A pasthrough query request for a store that can execute a subset of what is allowed by crates/sui-indexer-alt-graphql/schema.graphql::multiGetObjects. That query is likely to allow more than what is needed by the replay tool, and it’s pretty trivial in its usage.

TransactionStore

A TransactionStore has to be able to retrieve transaction data for a given digest. To replay a transaction the data in input to sui_execution::executor::Executor::execute_transaction_to_effects has to be provided. Some of that data is not provided by the user. It is naturally available at runtime on a live system and later saved in effects and in the context of a checkpoint.