sui_replay_2::replay_interface
Trait ObjectStore

pub trait ObjectStore {
    // Required method
    fn get_objects(
        &self,
        keys: &[ObjectKey],
    ) -> Result<Vec<Option<Object>>, Error>;
}

A pasthrough query request for a store that can execute a subset of what is allowed by crates/sui-indexer-alt-graphql/schema.graphql::multiGetObjects. That query is likely to allow more than what is needed by the replay tool, and it’s pretty trivial in its usage.

Required Methods

fn get_objects(&self, keys: &[ObjectKey]) -> Result<Vec<Option<Object>>, Error>