sui_replay_2::replay_interface
Trait SetupStore

pub trait SetupStore {
    // Required method
    fn setup(&self, chain_id: Option<String>) -> Result<Option<String>, Error>;
}

A trait to set up the data store. This is used to setup internal state of the data store before starting the replay. At the moment is exclusively used by the FileSystemStore to map network to chain id.

Required Methods

fn setup(&self, chain_id: Option<String>) -> Result<Option<String>, Error>

Set up the data store. Returns the chain identifier if available, or None if not available. When chain_id is None this is a no-op, unless the given data store has a way to fetch the chain id from the network. When chain_id is Some(chain_id) the given data store should override the map from network to chain id if it has one. That is a meaningful operation only for the FileSystemStore.